Add support for shiny and shiny_prerendered runtimes.
If rcloud.shiny is available rcloud.flexdashboard will support shiny and shiny_prerendered runtimes.
shiny integration is implemented by overriden functions from rmarkdown so they use rcloud.shiny implementations of shiny startup functions.

Temporary resources created rmarkdown are removed when RCloud session is closed.

FIX #20

+++ R/render.R
@@ -4,34 +4,50 @@ renderFlexDashboard <- function(id, version = NULL) {
     stop("Notebook \"", URLencode(id, TRUE), "\" does not exist or has not been published")
   tmp_dir <- tempfile("flexdashboard")
   dir.create(tmp_dir)
-  on.exit(unlink(tmp_dir), add = TRUE)
   
-  tmp <- tempfile(tmpdir = tmp_dir, fileext = ".Rmd")
-  exportRmd(id, version, file = tmp)
+  tmpRmd <- tempfile(tmpdir = tmp_dir, fileext = ".Rmd")
+  exportRmd(id, version, file = tmpRmd)
   exportCss(id, version, tmp_dir)
   
-  tmp2 <- tempfile(fileext = ".html", tmpdir = tmp_dir)
+  front_matter <- rmarkdown:::yaml_front_matter(tmpRmd)
+  
+  
   knitr:::knit_hooks$set(eval = progress_hook)
   on.exit(knitr:::knit_hooks$set(eval = NULL), add = TRUE)
-  render(
-    input = tmp,
-    output_file = tmp2
-  )
-  contents <- paste(readLines(tmp2), collapse = "\n")
-
-  caps$render(
-    "#rcloud-flexdashboard", 
-    gsub("\"", "&quot;", contents)
-  )
+  
+  if('runtime' %in% names(front_matter) && front_matter$runtime %in% c('shiny', 'shiny_prerendered')) {
+    addSessionCloseCallback(rmDirCallback(tmp_dir))
+    if (requireNamespace("rcloud.shiny", quietly = TRUE)) {
+      rmarkdown.run.override(tmpRmd, 
+                             auto_reload = FALSE,
+                             renderer = function(url) {   
+                               caps$renderShinyUrl("#rcloud-flexdashboard", url)
+                               rcw.result(run = function(...) { }, body = "")
+                             })
+    } else {
+      stop("flexdashboards that use shiny runtime require rcloud.shiny package, please install it and try again.")
+    }
+  } else {
+    tmpHtml <- tempfile(fileext = ".html", tmpdir = tmp_dir)
+    on.exit(unlink(tmp_dir), add = TRUE)
+    render(
+      input = tmpRmd, 
+      output_file = basename(tmpHtml)
+    )
+    contents <- paste(readLines(tmpHtml), collapse = "\n")
+    caps$render(
+      "#rcloud-flexdashboard", 
+      gsub("\"", "&quot;", contents)
+    )
+  }
 
   invisible()
 }
 
-
 exportCss <- function(notebook_id, version = NULL, tmp_dir) {
     res <- rcloud.support::rcloud.get.notebook(notebook_id, version)
     
-    if (! res$ok) {
+    if (!res$ok) {
       return(NULL)
     }
     
@@ -45,9 +61,30 @@ exportCss <- function(notebook_id, version = NULL, tmp_dir) {
     });
 }
 
+rmDirCallback <- function(dir.path) {
+  directory <- dir.path
+  function() {
+    if(!is.null(directory) && file.exists(directory)) {
+      unlink(directory, recursive = TRUE)
+    }
+  }
+}
+
+addSessionCloseCallback <- function(FUN) {
+  f <- .GlobalEnv$.Rserve.done
+  .GlobalEnv$.Rserve.done <- function(...) {
+    FUN()
+    if(is.function(f)) {
+      f(...)
+    }
+  }
+  
+}
+
 format_cell <- function(cell) {
   paste0(cell$content, "\n")
 }
+
 progress <- function(status) {
   if(exists("caps")) {
     caps$progress(
@@ -63,4 +100,3 @@ progress_hook <- function(before, options, envir) {
   progress(options$label)
   options
 }
-
